ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# NLP
spacy==3.7.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# AI
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
